nicely separate base OcppConnectionComponent trait and v1x specific stuff
